more refactor

+++ Section015_029-MERN/server/api/controllers/TaskController.js
@@ -2,19 +2,21 @@ const TaskServices = require("../../services/TaskServices");
 const { taskResponses, projectResponses, errorResponse } = require('../../common/res');
 const { validateRequest } = require("../../common/validations");
 const ProjectServices = require("../../services/ProjectServices");
+const ServiceHelper = require("../../common/utils/ServiceHelper");
 
 
 class TaskController {
     constructor() {
         this.taskServices = new TaskServices();
-        this.projectServices = new ProjectServices()
+        this.projectServices = new ProjectServices();
+        this.serviceHelper = new ServiceHelper(this.projectServices, this.taskServices);
     }
 
     async Create(request, response) {
         try {
             validateRequest(request);
 
-            if(!(await this.FindProjectAndOwnership(request, response)))
+            if(!(await this.serviceHelper.FindProjectAndOwnership(request, response)))
                 return;
 
             const task = await this.taskServices.CreateTask(request.body)
@@ -30,7 +32,7 @@ class TaskController {
         try {
             validateRequest(request);
 
-            if(!(await this.FindProjectAndOwnership(request, response)))
+            if(!(await this.serviceHelper.FindProjectAndOwnership(request, response)))
                 return;
             
             const tasks = await this.taskServices.FindProjectTasks(request.body.projectId);
@@ -46,7 +48,7 @@ class TaskController {
         try {
             validateRequest(request);
 
-            if(!(await this.FindProjectOwnershipAndFindTask(request, response)))
+            if(!(await this.serviceHelper.FindProjectOwnershipAndFindTask(request, response)))
                 return;
 
             const newTask = this.taskServices.ConstructNewTask(request.body);
@@ -62,6 +64,9 @@ class TaskController {
     async Delete(request, response) {
         try {
             validateRequest(request);
+            
+            if(!(await this.serviceHelper.FindProjectOwnershipAndFindTask(request, response)))
+                return;
 
             await this.taskServices.DeleteTask(request.params.id);
             taskResponses.taskDeleted(response);
@@ -71,34 +76,6 @@ class TaskController {
         catch (error) {
             errorResponse(response, 'Error while creating the task', error);
         }
-    }
-
-    async FindProjectAndOwnership(request, response) {
-        const project = await this.projectServices.FindProjectById(request.body.projectId)   
-        if(!project) {
-            projectResponses.projectNotFound(response);
-            return false;
-        }
-                
-        if(!this.projectServices.ValidateUserProjectOwnership(request.user.id, project.creator)) {
-            projectResponses.projectNotOwned(response);
-            return false;
-        }
-        
-        return true;
-    }
-
-    async FindProjectOwnershipAndFindTask(request, response) {
-        if(!(await this.FindProjectAndOwnership(request, response))) {
-            return false;
-        }
-
-        if(!(await this.taskServices.FindTaskById(request.params.id))) {
-            taskResponses.taskNotFound(response);
-            return false;
-        }
-
-        return true;
     }
 }
